Adapt workflow to human MetaT paired reads data

+++ metatranscriptomics/pyscripts_edited/6_BWA_Gene_Map.py
@@ -35,7 +35,7 @@ for contig in contig2read_map_full:
 gene2read_map = {}
 mapped_reads = set()
 prev_mapping_count = 0
-for x in range((len(sys.argv) - 5) / 3):
+for x in range(int((len(sys.argv) - 5) / 3)):
     read_file = sys.argv[3 * x + 5]
     read_seqs = SeqIO.index(read_file, os.path.splitext(read_file)[1][1:])
     BWA_sam_file = sys.argv[3 * x + 6]
